Frontend/Backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
pypdf